requirements.txt
langchain-ollama
langchain-chroma
langchain-community
faiss-cpu